Renamed info to infofile
