+++ infofile/api.go
@@ -1,4 +1,4 @@
-package info
+package infofile
 
 import (
 	"fmt"
