+++ infofile/api_fs_test.go
@@ -1,9 +1,9 @@
-package info
+package infofile
 
 import (
 	"testing"
 
-	"github.com/jwaldrip/treex/info/internal/testutil"
+	"github.com/jwaldrip/treex/infofile/internal/testutil"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
