+++ infofile/api_unit_test.go
@@ -1,4 +1,4 @@
-package info
+package infofile
 
 import (
 	"testing"
