+++ infofile/cmd/commands/clean.go
