+++ infofile/cmd/commands/remove.go
