+++ treex/plugins/infofile/infofile.go
@@ -1,5 +1,5 @@
-// Package info provides a plugin for annotation-based file filtering using .info files
-package info
+// Package infofile provides a plugin for annotation-based file filtering using .info files
+package infofile
 
 import (
 	"log"
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/jwaldrip/treex/info"
+	infofile "github.com/jwaldrip/treex/infofile"
 	"github.com/jwaldrip/treex/treex/plugins"
 	"github.com/jwaldrip/treex/treex/types"
 	"github.com/spf13/afero"
@@ -90,7 +90,7 @@ func (p *InfoPlugin) ProcessRoot(fs afero.Fs, rootPath string) (*plugins.Result,
 	result.Categories["annotated"] = make([]string, 0)
 
 	// Create an InfoAPI to parse .info files in this root
-	api := info.NewInfoAPI(fs)
+	api := infofile.NewInfoAPI(fs)
 	annotations, err := api.Gather(rootPath)
 	if err != nil {
 		// If we can't collect annotations, return empty result (not an error)
@@ -128,14 +128,14 @@ func (p *InfoPlugin) ProcessRoot(fs afero.Fs, rootPath string) (*plugins.Result,
 func (p *InfoPlugin) GetAnnotationDetails(fs afero.Fs, rootPath string) (map[string]interface{}, error) {
 	details := make(map[string]interface{})
 
-	api := info.NewInfoAPI(fs)
+	api := infofile.NewInfoAPI(fs)
 	annotations, err := api.Gather(rootPath)
 	if err != nil {
 		return details, err
 	}
 
 	// Group annotations by .info file (only winning annotations are included)
-	byInfoFile := make(map[string][]info.Annotation)
+	byInfoFile := make(map[string][]infofile.Annotation)
 	for _, annotation := range annotations {
 		infoFile := annotation.InfoFile
 		byInfoFile[infoFile] = append(byInfoFile[infoFile], annotation)
@@ -183,7 +183,7 @@ func (p *InfoPlugin) EnrichNode(fs afero.Fs, node *types.Node) error {
 	}
 
 	// Use the InfoAPI to find annotation for this specific path
-	api := info.NewInfoAPI(fs)
+	api := infofile.NewInfoAPI(fs)
 
 	// Try to find annotation starting from the node's directory
 	searchPath := "."
@@ -224,7 +224,7 @@ func (p *InfoPlugin) EnrichData(fs afero.Fs, rootPath string, filePaths []string
 
 	// Check if we have cached annotations from the filtering phase
 	if cachedAnnotations, exists := cache["annotations"]; exists {
-		if annotations, ok := cachedAnnotations.(map[string]info.Annotation); ok {
+		if annotations, ok := cachedAnnotations.(map[string]infofile.Annotation); ok {
 			// Use cached annotations for efficient enrichment
 			for _, filePath := range filePaths {
 				// Look for annotation for this specific file
@@ -259,7 +259,7 @@ func (p *InfoPlugin) EnrichData(fs afero.Fs, rootPath string, filePaths []string
 		}
 	} else {
 		// No cached data available, gather annotations fresh
-		api := info.NewInfoAPI(fs)
+		api := infofile.NewInfoAPI(fs)
 		annotations, err := api.Gather(rootPath)
 		if err != nil {
 			// If we can't gather annotations, return empty map (not an error)
@@ -305,7 +305,7 @@ func (p *InfoPlugin) EnrichNodeWithCache(fs afero.Fs, node *types.Node, pluginRe
 		}
 
 		// Type assert to get the annotations map
-		annotations, ok := cachedAnnotations.(map[string]info.Annotation)
+		annotations, ok := cachedAnnotations.(map[string]infofile.Annotation)
 		if !ok {
 			continue
 		}
